wt-status: read rebase todolist with strbuf_getline()
In read_rebase_todolist() every line is, after reading, checked for a
comment_line_char. After that it is trimmed via strbuf_trim().
Assuming we do never expect a CR as comment_line_char,
strbuf_getline_lf() can be safely replaced by its CRLF counterpart.

The current code checks if the line begins with a comment line char
(typically '#') before trimming, which is consistent with the way how
commands like 'git commit' prepares commented lines in that this does
not treat a line that begins with whitespaces and '#' as commented out.
We however made a mistake in the parser of "git rebase -i" long
time ago to allow such a line to be treated as a comment, and made an
exception with 1db168ee (rebase-i: loosen over-eager check_bad_cmd
check, 2015-10-01).  It probably is a good idea to match that
exception by swapping the order of comment check and trimming in this
codepath as well.
--> this is obsolete by the latter comment!

These files are already preprocessed with git stripspace when stored
in $GIT_DIR thus neither CR nor whitespaces are to be expected.

Helped-by: Junio C Hamano <gitster@pobox.com>